Update Window Resynthesis (#503)
* feature(resub): implement the 2-resub with 3 unate

* feature(resub): implementation of OR-AND and AND-OR 2-resub

* update(resub): add commment about the OR balance and AND balance in the resub implementation

* feature(resub): implement theOR-2OR, OR-2AND, AND-2OR, AND-2AND of 3-resub

* feature(test): implement a graph based test for 0,1,2,3-resub

* update(mig-resub): modify the experimental to use MIG

* feature(mig-resub): implement2-resub of MIG

* fupdate(mig-resub): fix the binate divisor collecting method

* update(resub): reset previous changes to the prev code

* update(resub): clean up all the experiment log

* update(resubfix the mis-delete in prev cleanup

* upd(resub): use const reference to avoid copy

* update(resub): modify some comment

* update(mig-resub) : clean up experiment logs

* Fix for Windows

* clang-format

+++ include/mockturtle/algorithms/experimental/sim_resub.hpp
@@ -62,7 +62,7 @@ struct breadth_first_windowing_params
   uint32_t max_divisors{50};
 
   /*! \brief Maximum number of TFI nodes to collect. */
-  uint32_t max_tfi{max_divisors * 0.5};
+  uint32_t max_tfi{uint32_t(max_divisors * 0.5)};
 
   /*! \brief Maximum number of nodes added by resubstitution. */
   uint32_t max_inserts{std::numeric_limits<uint32_t>::max()};
